refactoring

+++ PurseApp/Controllers/PurseController.cs
@@ -4,11 +4,12 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using PurseApp.Models;
+using PurseApp.Models.Dto;
 using PurseApp.Repositories;
 
 namespace PurseApp.Controllers
 {
-    [Route("api/[controller]")]
+    [Route("api/{userId}/[controller]")]
     [ApiController]
     public class PurseController : ControllerBase
     {
@@ -21,7 +22,7 @@ namespace PurseApp.Controllers
             _purseRepository = purseRepository;
         }
         
-        [HttpGet("purse/{userId}")]
+        [HttpGet]
         public async Task<ActionResult<IEnumerable<Purse>>> GetPurses(Guid userId)
         {
             if (! await _userRepository.IsUserExists(userId))
@@ -29,19 +30,21 @@ namespace PurseApp.Controllers
             return Ok(await _purseRepository.GetPurses(userId));
         }
         
-        [HttpGet("purse/accounts/{userId}")]
-        public async Task<ActionResult<IEnumerable<(string,string,string)>>> GetPursesBalance(Guid userId)
+        [HttpGet("accounts")]
+        public async Task<ActionResult<IEnumerable<PurseBalanceDto>>> GetPursesBalance(Guid userId)
         {
             if (! await _userRepository.IsUserExists(userId))
                 return NotFound("Пользователь не существует");
             var purses = (await _purseRepository.GetPurses(userId)).SelectMany(s => s.Accounts)
-                .Select(s => new {accountname = s.Name, s.Balance, currencyname = s.Currency.Name});
+                .Select(s => new PurseBalanceDto {AccountName = s.Name, Balance = s.Balance, Currency = s.Currency.Name});
             return Ok(purses);
         }
 
-        [HttpDelete("purse/{purseId}")]
-        public async Task<IActionResult> DeletePurse(Guid purseId)
+        [HttpDelete("{purseId}")]
+        public async Task<IActionResult> DeletePurse(Guid userId, Guid purseId)
         {
+            if (!await _userRepository.IsUserExists(userId))
+                return NotFound();
             await _purseRepository.RemovePurse(purseId);
             return Ok();
         }
